requirements.txt
streamlit==1.28.0
spacy==3.7.0
PyMuPDF==1.22.5
python-docx==0.8.11
pandas==2.1.1
